@@ -12,55 +12,55 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@apollo/client": "^3.5.8",
-    "@babel/core": "^7.17.2",
+    "@apollo/client": "^3.5.10",
+    "@babel/core": "^7.17.9",
     "@babel/eslint-parser": "^7.17.0",
     "@babel/plugin-proposal-class-properties": "^7.16.7",
     "@babel/plugin-transform-runtime": "^7.17.0",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
-    "@babel/register": "^7.17.0",
-    "@babel/runtime": "^7.17.2",
-    "autoprefixer": "^10.4.2",
-    "axios": "^0.25.0",
+    "@babel/register": "^7.17.7",
+    "@babel/runtime": "^7.17.9",
+    "autoprefixer": "^10.4.4",
+    "axios": "^0.26.1",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.3",
-    "body-parser": "^1.19.1",
+    "babel-loader": "^8.2.4",
+    "body-parser": "^1.20.0",
     "compression-webpack-plugin": "^9.2.0",
     "connect-mongo": "^4.6.0",
     "copy-webpack-plugin": "^10.2.4",
-    "css-loader": "^6.6.0",
+    "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^3.4.1",
-    "eslint": "^8.8.0",
+    "eslint": "^8.13.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "express": "^4.17.2",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
+    "express": "^4.17.3",
     "express-graphql": "^0.12.0",
     "express-session": "^1.17.2",
     "graphql": "^16.3.0",
     "html-webpack-plugin": "^5.5.0",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.5.3",
-    "mongoose": "6.2.1",
+    "mini-css-extract-plugin": "^2.6.0",
+    "mongoose": "6.2.11",
     "nodemon": "2.0.15",
     "passport": "^0.5.2",
     "passport-local": "^1.0.0",
     "postcss-loader": "^6.2.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router": "^6.2.1",
-    "react-router-dom": "^6.2.1",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "remove-files-webpack-plugin": "^1.5.0",
-    "sass": "^1.49.7",
-    "sass-loader": "^12.4.0",
+    "sass": "^1.50.0",
+    "sass-loader": "^12.6.0",
     "sass-resources-loader": "^2.2.4",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
-    "webpack": "^5.68.0",
+    "webpack": "^5.72.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-dev-middleware": "^5.3.1",
     "webpack-hot-middleware": "^2.25.1"
